page restricted

+++ base/views.py
@@ -1,6 +1,12 @@
 from django.http import HttpResponse
 from django.shortcuts import render, redirect
-from .models import Room
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.models import User
+from django.db.models import Q
+from .models import Room, Topic
 from .forms import RoomForm
 
 # Create your views here.
@@ -11,9 +17,51 @@ from .forms import RoomForm
 #     {'id':3, 'name': 'Learn C++'},
 # ]
 
+def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+
+        try: #if user exist
+            user = User.objects.get(username=username)
+        except: #if user doesn't exist
+            messages.add_message(request, messages.INFO, "User Not Found!")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.add_message(request, messages.INFO, "Username or Password doesnot exist!")
+
+
+
+    context = {}
+    return render(request, "base/login_register.html", context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+
 def home(request):
-    rooms = Room.objects.all()
-    context = {'rooms':rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q) |
+        Q(description__icontains=q)
+        )
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+    context = {'rooms':rooms, 'topics':topics, 'room_count':room_count}
     return render(request, "base/home.html", context)
 
 def room(request, pk):
@@ -21,6 +69,7 @@ def room(request, pk):
     context = {'room':room}
     return render(request, "base/room.html",context)
 
+@login_required(login_url='/login')
 def create_room(request):
     form = RoomForm
     if request.method == "POST":
@@ -32,9 +81,13 @@ def create_room(request):
     context = {'form':form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url='/login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room) #pre-fill form
+
+    if request.user != room.host:
+        return HttpResponse('You are not authorized to perform this action!!')
     
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -46,9 +99,16 @@ def update_room(request, pk):
     context = {'form': form}
     return render(request, "base/room_form.html",context)
 
+@login_required(login_url='/login')
 def delete_room(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not authorized to perform this action!!')
+    
+
     if request.method=="POST":
         room.delete()
         return redirect('home')
     return render(request, "base/delete.html")
+
